+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 This is the changelog of our "JTL WooCommerce Connector".
 
 ## Unreleased
+- CO-3503 - add telemetry config flag and event DTOs (steps 1-2)
 - CO-3161 - consider different customer languages
 - CO-3308 - fix seo url changes on update
 - CO-3262 - update wpml plugin name